LayoutTests/fast/css/box-sizing-backwards-compat-prefix.html

<!DOCTYPE html>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="test-prefixed-sizing" style="-webkit-box-sizing: border-box;"></div>
<div id="test-prefixless-sizing" style="box-sizing: border-box;"></div>
<div id="console"></div>
<script>
description("Testing whether box-sizing works with and withou the -webkit-prefix (bug 36713).");

var prefixStyle = document.getElementById('test-prefixed-sizing').style;
shouldBeEqualToString("prefixStyle.getPropertyValue('box-sizing')", "border-box");
shouldBeEqualToString("prefixStyle.getPropertyValue('-webkit-box-sizing')", "border-box");
shouldBeEqualToString("prefixStyle.boxSizing", "border-box");
shouldBeEqualToString("prefixStyle.WebkitBoxSizing", "border-box");
shouldBeEqualToString("prefixStyle.webkitBoxSizing", "border-box");

var prefixLessStyle = document.getElementById('test-prefixless-sizing').style;
shouldBeEqualToString("prefixLessStyle.getPropertyValue('box-sizing')", "border-box");
shouldBeEqualToString("prefixLessStyle.getPropertyValue('-webkit-box-sizing')", "border-box");
shouldBeEqualToString("prefixLessStyle.boxSizing", "border-box");
shouldBeEqualToString("prefixLessStyle.WebkitBoxSizing", "border-box");
shouldBeEqualToString("prefixLessStyle.webkitBoxSizing", "border-box");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
