third_party/WebKit/LayoutTests/fast/css/box-sizing-backwards-compat-prefix-expected.txt

Testing whether box-sizing works with and withou the -webkit-prefix (bug 36713).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS prefixStyle.getPropertyValue('box-sizing') is "border-box"
PASS prefixStyle.getPropertyValue('-webkit-box-sizing') is "border-box"
PASS prefixStyle.boxSizing is "border-box"
PASS prefixStyle.WebkitBoxSizing is "border-box"
PASS prefixStyle.webkitBoxSizing is "border-box"
PASS prefixLessStyle.getPropertyValue('box-sizing') is "border-box"
PASS prefixLessStyle.getPropertyValue('-webkit-box-sizing') is "border-box"
PASS prefixLessStyle.boxSizing is "border-box"
PASS prefixLessStyle.WebkitBoxSizing is "border-box"
PASS prefixLessStyle.webkitBoxSizing is "border-box"
PASS successfullyParsed is true

TEST COMPLETE
